requirements.txt
blessings==1.6
certifi==2017.7.27.1
chardet==3.0.4
construct==2.5.3
curtsies==0.2.11
Django==1.11.6
django-haystack==2.6.1
django-pagedown==0.1.3
future==0.16.0
greenlet==0.4.12
idna==2.6
jieba==0.39
Markdown==2.6.9
pefile==2017.11.5
Pygments==2.2.0
PyMySQL==0.7.11
python-ptrace==0.9.3
pytz==2017.2
requests==2.18.4
six==1.11.0
urllib3==1.22
uWSGI==2.0.17
wcwidth==0.1.7
Whoosh==2.7.4